example-django-app/requirements/base.txt
Django>=2.1.0,<2.2.0
djangorestframework>=3.10.0
djangorestframework-jsonapi>=2.8.0
django-filter>=2.2.0
